+++ src/psychic-app-websockets/index.ts
@@ -53,6 +53,12 @@ export default class PsychicAppWebsockets {
     return (psychicWebsocketsApp.psychicApp.constructor as typeof PsychicApp).log(...args)
   }
 
+  public static logWithLevel(...args: Parameters<typeof PsychicApp.logWithLevel>) {
+    const psychicWebsocketsApp = this.getOrFail()
+
+    return (psychicWebsocketsApp.psychicApp.constructor as typeof PsychicApp).logWithLevel(...args)
+  }
+
   constructor(psychicApp: PsychicApp) {
     this.psychicApp = psychicApp
   }